+++ hashfs/__meta__.py
@@ -6,7 +6,7 @@ __title__ = 'hashfs'
 __summary__ = 'A content-addressable file management system.'
 __url__ = 'https://github.com/dgilland/hashfs'
 
-__version__ = '0.0.1'
+__version__ = '0.1.0'
 
 __install_requires__ = []
 __tests_require__ = ['tox']