+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-vx.x.x (xxxx-xx-xx)
+v0.1.0 (2015-05-28)
 -------------------
 
 - Add ``HashFS.get()`` method for retrieving a reader object given a file digest or path.